Rename sql/src/plan/expr.rs to hir.rs
